moved IDs to Long for Lab 9

+++ Lab-09-solution/src/main/java/com/petclinic/core/OwnerRepository.java
@@ -5,8 +5,9 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.Optional;
 
-public interface OwnerRepository extends JpaRepository<Owner, Integer> {
+public interface OwnerRepository extends JpaRepository<Owner, Long> {
     @Query("SELECT o FROM Owner o LEFT JOIN FETCH o.pets WHERE o.firstName = :firstName")
     public Optional<Owner> findByFirstNameWithPets(String firstName);
+
     public Optional<Owner> findByFirstName(String firstName);
 }
